Extract ObjectCacheKey to its own file from ObjectQuery.ts
