+++ packages/client/src/observable/internal/object/ObjectsHelper.ts
@@ -19,7 +19,8 @@ import type { ObjectPayload } from "../../ObjectPayload.js";
 import type { ObserveObjectOptions } from "../../ObservableClient.js";
 import type { Observer } from "../../ObservableClient/common.js";
 import { AbstractHelper } from "../AbstractHelper.js";
-import { type ObjectCacheKey, ObjectQuery } from "../ObjectQuery.js";
+import { type ObjectCacheKey } from "../ObjectCacheKey.js";
+import { ObjectQuery } from "../ObjectQuery.js";
 import type { QuerySubscription } from "../QuerySubscription.js";
 
 export class ObjectsHelper extends AbstractHelper<
